requirements.txt
absl-py==1.4.0
attrs==22.2.0
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cosine-annealing-warmup==2.0
cycler==0.11.0
filelock==3.10.0
fonttools==4.38.0
freetype-py==2.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.0.0
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
kiwisolver==1.4.4
lmdb==1.4.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdurl==0.1.2
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
pyrender==0.1.45
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
seaborn==0.12.2
sentencepiece==0.1.97
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tokenizers==0.10.3
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.65.0
transformers==4.12.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
